+++ seqspec/seqspec_find.py
@@ -68,12 +68,12 @@ def validate_find_args(parser, args):
 
 
 def run_find(spec: Assay, modality: str, region_id: str):
-    m = spec.get_modality(modality)
+    m = spec.get_libspec(modality)
     regions = m.get_region_by_id(region_id)
     return regions
 
 
 def run_find_by_type(spec: Assay, modality: str, region_type: str):
-    m = spec.get_modality(modality)
+    m = spec.get_libspec(modality)
     regions = m.get_region_by_type(region_type)
     return regions